pyproject.toml
[tool.poetry]
name = "realtime-transcription-chatbot"
version = "0.1.0"
description = "Realtime transcription chatbot"
authors = ["Nikhil Cheerla"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10.0"
openai = "^1.44.0"
elevenlabs = "^1.9.0"
assemblyai = {extras = ["extras"], version = "^0.34.0"}
python-dotenv = "^1.0.1"
pyaudio = "^0.2.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"